server/requirements.txt
Flask>=2.0
flask-cors
SQLAlchemy
PyMySQL
python-dotenv
pytest
requests
